Histogram Plots Comparision of two Algorithms
We will use this plot to visualize which ad version is mostly selected or clicked online. It is very clear, In this case it is ad version 5 is mostly clicked. Comparing UCB and Thompson_Sampling Plots We can clearly See Thompson is Performing better in giving rewards and selecting the right Ad_version. In Both Cases Ad_version is most likely Ad_version But We can clearly see the performances of both algorithms.